refactor: Iconフォルダーをcomponents以下に移動した
