+++ packages/zenn-embed-components/src/components/Icons/GithubStarIcon.tsx
